# PERF/REF: MultiIndex.copy